refactor(cordis): move the backend out of the package __init__
reef/train/cordis_backend/__init__.py held 857 lines of implementation. Every
other package __init__ in reef is a docstring and re-exports, and the sibling
slime_backend keeps its backend in backend.py, so the one place a reader looks
for the package's contract was the one place that hid it.

Hosting the implementation there also forced an import cycle workaround:
recipe.py imports CordisBackend from the package, so the package's own
submodule imports had to sit at the bottom of the file behind noqa: F401, under
two comment blocks that said nearly the same thing.

CordisBackend, HarnessCandidate and ScoreComparisonSelector move to backend.py.
recipe.py imports them from there, so __init__.py is a docstring, ordinary
top-of-file re-exports, and an __all__ it did not have before.

tests/reef_service/test_harness_recipe.py patches run_episode on the module
that resolves it, which is now backend.py.
